AI_Service/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
openai>=1.0.0
opencv-python
numpy
Pillow
scikit-learn
torch
torchvision
python-multipart